refactor: moved all new Media hooks to the plugins/core/files/server/methods, updated medaGallery to use the new Meteor media methods, media update and remove hooks added and working
