+++ imports/plugins/core/files/server/methods.js
@@ -1,20 +1,19 @@
 import { Meteor } from "meteor/meteor";
 import { check } from "meteor/check";
-import includes from "lodash/includes";
-import { Hooks } from "/server/api";
-import { Reaction } from "/server/api";
+import { Hooks, Reaction } from "/server/api";
 import { MediaRecords, Revisions } from "/lib/collections";
+import { Media } from "/imports/plugins/core/files/server";
 import { RevisionApi } from "/imports/plugins/core/revisions/lib/api";
 
 /**
  * @method updateMediaMetadata
  * @memberof media
- * @summary insert a new media record and add it to revision control.
- * @param {Object} fileRecord - document from file collection upload.
- * @return {String} - _id of the new inserted media record.
+ * @summary updates media record in revision control.
+ * @param {String} fileRecordId - _id of updated file record.
+ * @param {Object} metadata - metadata from updated media file.
+ * @return {Boolean}
  */
 export async function updateMediaMetadata(fileRecordId, metadata) {
-  console.log("Update Media Metadata", fileRecordId, metadata)
   check(fileRecordId, String);
   check(metadata, Object);
   if (RevisionApi.isRevisionControlEnabled()) {
@@ -63,51 +62,123 @@ export async function updateMediaMetadata(fileRecordId, metadata) {
   return true;
 }
 
+/**
+ * @method insertMedia
+ * @memberof media
+ * @summary insert a new media record and add it to revision control.
+ * @param {Object} fileRecord - document from file collection upload.
+ * @return {String} - _id of the new inserted media record.
+ */
+export async function insertMedia(fileRecord) {
+  check(fileRecord, Object);
+  const mediaRecord = await MediaRecords.insert(fileRecord);
 
-Meteor.methods({
-  /**
-   * updateMediaPriorities
-   * @summary sorting media by array indexes
-   * @type {Method}
-   * @param {String[]} sortedMediaIDs
-   * @return {Boolean} true
-   */
-  "media/updatePriorities"(sortedMediaIDs) {
-    check(sortedMediaIDs, [String]);
-
-    if (!Reaction.hasPermission("createProduct")) {
-      throw new Meteor.Error("access-denied", "Access Denied");
-    }
-
-    // Check to be sure product linked with each media belongs to the current user's current shop
-    const shopId = Reaction.getShopId();
-
-    const sortedMediaRecords = MediaRecords.find({
-      _id: { $in: sortedMediaIDs }
-    }).fetch();
-
-    sortedMediaRecords.forEach((mediaRecord) => {
-      if (!mediaRecord.metadata || mediaRecord.metadata.shopId !== shopId) {
-        throw new Meteor.Error("access-denied", `Access Denied. No access to shop ${mediaRecord.metadata.shopId}`);
-      }
-    });
-
-    if (sortedMediaRecords.length !== sortedMediaIDs.length) {
-      throw new Meteor.Error("not-found", "At least one ID in sortedMediaIDs does not exist");
-    }
-
-    sortedMediaIDs.forEach((_id, index) => {
-      MediaRecords.update({
-        _id
-      }, {
-        $set: {
-          "metadata.priority": index
+  if (RevisionApi.isRevisionControlEnabled() && fileRecord.metadata.workflow !== "published") {
+    if (fileRecord.metadata.productId) {
+      const revisionMetadata = Object.assign({}, fileRecord.metadata);
+      revisionMetadata.workflow = "published";
+      Revisions.insert({
+        documentId: mediaRecord,
+        documentData: revisionMetadata,
+        documentType: "image",
+        parentDocument: fileRecord.metadata.productId,
+        changeType: "insert",
+        workflow: {
+          status: "revision/update"
         }
       });
-      const { metadata } = MediaRecords.findOne({ _id });
-      updateMediaMetadata(_id, metadata);
-    });
+      fileRecord.metadata.workflow = "unpublished";
+    } else {
+      fileRecord.metadata.workflow = "published";
+    }
+  }
 
+  return mediaRecord;
+}
+
+/**
+ * @method removeMedia
+ * @memberof media
+ * @summary removes media file and updates record in revision control.
+ * @param {String} fileRecordId - _id of updated file record.
+ * @param {Object} metadata - metadata from updated media file.
+ * @return {Boolean}
+ */
+export async function removeMedia(fileRecordId) {
+  check(fileRecordId, String);
+  const { metadata } = MediaRecords.findOne({ _id: fileRecordId });
+  Media.remove(fileRecordId);
+  console.log("removeMedia", metadata)
+  if (RevisionApi.isRevisionControlEnabled() && metadata.workflow && metadata.workflow === "unpublished") {
+    Revisions.remove({
+      documentId: fileRecordId
+    });
     return true;
   }
+  if (metadata.productId) {
+    Revisions.insert({
+      documentId: fileRecordId,
+      documentData: metadata,
+      documentType: "image",
+      parentDocument: metadata.productId,
+      changeType: "remove",
+      workflow: {
+        status: "revision/update"
+      }
+    });
+    return false; // prevent actual deletion of image. This also stops other hooks from running :/
+  }
+  return true;
+}
+
+/**
+ * updateMediaPriorities
+ * @summary sorting media by array indexes
+ * @type {Method}
+ * @param {String[]} sortedMediaIDs
+ * @return {Boolean} true
+ */
+export function updateMediaPriorities(sortedMediaIDs) {
+  check(sortedMediaIDs, [String]);
+
+  if (!Reaction.hasPermission("createProduct")) {
+    throw new Meteor.Error("access-denied", "Access Denied");
+  }
+
+  // Check to be sure product linked with each media belongs to the current user's current shop
+  const shopId = Reaction.getShopId();
+
+  const sortedMediaRecords = MediaRecords.find({
+    _id: { $in: sortedMediaIDs }
+  }).fetch();
+
+  sortedMediaRecords.forEach((mediaRecord) => {
+    if (!mediaRecord.metadata || mediaRecord.metadata.shopId !== shopId) {
+      throw new Meteor.Error("access-denied", `Access Denied. No access to shop ${mediaRecord.metadata.shopId}`);
+    }
+  });
+
+  if (sortedMediaRecords.length !== sortedMediaIDs.length) {
+    throw new Meteor.Error("not-found", "At least one ID in sortedMediaIDs does not exist");
+  }
+
+  sortedMediaIDs.forEach((_id, index) => {
+    MediaRecords.update({
+      _id
+    }, {
+      $set: {
+        "metadata.priority": index
+      }
+    });
+    const { metadata } = MediaRecords.findOne({ _id });
+    updateMediaMetadata(_id, metadata);
+  });
+
+  return true;
+}
+
+Meteor.methods({
+  "media/insert": insertMedia,
+  "media/updatePriorities": updateMediaPriorities,
+  "media/remove": removeMedia
 });
