+++ imports/plugins/core/revisions/server/hooks.js
@@ -270,33 +270,33 @@ export const ProductRevision = {
 //   return true;
 // });
 
-MediaRecords.before.remove((userId, media) => {
-  if (RevisionApi.isRevisionControlEnabled() === false) {
-    return true;
-  }
+// MediaRecords.before.remove((userId, media) => {
+//   if (RevisionApi.isRevisionControlEnabled() === false) {
+//     return true;
+//   }
 
-  // if the media is unpublished, then go ahead and just delete it
-  if (media.metadata.workflow && media.metadata.workflow === "unpublished") {
-    Revisions.remove({
-      documentId: media._id
-    });
-    return true;
-  }
-  if (media.metadata.productId) {
-    Revisions.insert({
-      documentId: media._id,
-      documentData: media.metadata,
-      documentType: "image",
-      parentDocument: media.metadata.productId,
-      changeType: "remove",
-      workflow: {
-        status: "revision/update"
-      }
-    });
-    return false; // prevent actual deletion of image. This also stops other hooks from running :/
-  }
-  return true;
-});
+//   // if the media is unpublished, then go ahead and just delete it
+//   if (media.metadata.workflow && media.metadata.workflow === "unpublished") {
+//     Revisions.remove({
+//       documentId: media._id
+//     });
+//     return true;
+//   }
+//   if (media.metadata.productId) {
+//     Revisions.insert({
+//       documentId: media._id,
+//       documentData: media.metadata,
+//       documentType: "image",
+//       parentDocument: media.metadata.productId,
+//       changeType: "remove",
+//       workflow: {
+//         status: "revision/update"
+//       }
+//     });
+//     return false; // prevent actual deletion of image. This also stops other hooks from running :/
+//   }
+//   return true;
+// });
 
 /**
  * @function
